Rename: enum관련 패키지 이름 수정
- 모든 package 이름을 단수로 통일하기 위함
- enum -> enumeration
