+++ src/main/java/kappzzang/jeongsan/enumeration/Status.java
@@ -1,4 +1,4 @@
-package kappzzang.jeongsan.enums;
+package kappzzang.jeongsan.enumeration;
 
 public enum Status {
     ONGOING, PENDING, COMPLETED
